// src/app/model/device_basic.ts
import {DeviceValue} from "./device-value";

export class DeviceBasic {
  ALIAS: string = "";
  IDENTIFIER: string = "";
  LAST_CON: string = "";
  LAST_SOCKET: number = 0;
  LAST_IP: string = "";
  STATUS: Boolean = false;
  LATITUDE: number = 0;
  LONGITUDE: number = 0;
  PASSWORD: string ='';
  _id: string = ""

}